+++ app/Http/Controllers/ControlController.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace BristolSU\Playground\Http\Controllers;
+
+class ControlController extends Controller
+{
+
+    public function index()
+    {
+        return view('pages.control');
+    }
+
+}