+++ resources/js/control.js
@@ -1,0 +1,1 @@
+import '@bristol-su/control-management/src/main';