+++ resources/views/pages/control.blade.php
@@ -1,0 +1,12 @@
+@extends('bristolsu::base')
+
+@section('title', 'Control')
+
+@section('content')
+    <div id="control-management"></div>
+@endsection
+
+
+@push('scripts')
+    <script src="{{mix('js/control.js')}}"></script>
+@endpush