+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Fix js bug with loading old settings into the settings pane
+- Confirmation before module instance is deleted
+
+### Added
+- Added control to the playground
+- Assign Roles module
+
 ## [1.2.1] - (18/03/2020)
 
 ### Added